refactor: Simplify project structure
